+++ CHANGELOG.md
@@ -2,4 +2,9 @@
 
 ## Rugments 1.0.0 (2015-XX-XX)
 
-To be released!
+* Initial fork from [rouge](https://github.com/jneen/rouge)
+* Huge code cleanup (linting with rubocop)
+* Improved HTML formatter
+  * Changed `wrap` to `nowrap` which is `false` by default
+  * Ported `linenos`, `linenostart`, `lineanchors` and `anchorlinenos` from pygments
+  * See documentation here: http://pygments.org/docs/formatters/#HtmlFormatter